+++ Changelog.md
@@ -15,6 +15,7 @@ v2.11.0 (not published yet)
 - Fix remove confirmation on Lists showing first item name only even if multiple items are selected
 - Add NVEnc parameters --lut3d, --lut3d_interp, --dolby-vision-rpu, --dolby-vision-profile
 - Add VCEEnc parameter --thread-affinity
+- Add QSVEnc parameters --thread-affinity, --dhdr10-info, --dolby-vision-rpu, --dolby-vision-profile
 - Add x265 Dolby Vision Profile 8.4
 - Fix minor mkvextract demux issue ([#833](/../../issues/833))
 - Add demux support for DVBSUB subtitles ([#833](/../../issues/833))